fix: move amET locale to ethiopic package
