+++ .changeset/ethiopic-amet-locale.md
@@ -1,0 +1,6 @@
+---
+"@daypicker/ethiopic": patch
+"react-day-picker": patch
+---
+
+Move the `amET` locale export to `@daypicker/ethiopic`.
