+++ apps/website/versioned_docs/version-next/localization/ethiopic.mdx
@@ -18,12 +18,13 @@ import { DayPicker } from "@daypicker/ethiopic";
 
 ### Basic usage
 
-By default, the locale is `amET` with Ge'ez numerals.
+By default, the locale is `amET` with Ge'ez numerals. The `amET` locale is
+owned and exported by `@daypicker/ethiopic`.
 
 ```tsx
-import { DayPicker } from "@daypicker/ethiopic";
+import { DayPicker, amET } from "@daypicker/ethiopic";
 
-<DayPicker />;
+<DayPicker locale={amET} />;
 ```
 
 <BrowserWindow sourceUrl="https://github.com/gpbl/react-day-picker/blob/release/v10-next/examples/Ethiopic.tsx">
