+++ apps/website/versioned_docs/version-next/upgrading.mdx
@@ -140,6 +140,8 @@ These imports were removed in v10:
 - `react-day-picker/hebrew` - use `@daypicker/hebrew`
 - `react-day-picker/hijri` - use `@daypicker/hijri`
 - `react-day-picker/persian` - use `@daypicker/persian`
+- `amET` from `react-day-picker/locale` - use `amET` from
+  `@daypicker/ethiopic`
 - `enUSJalali` from `react-day-picker/locale` - use `enUSJalali` from
   `@daypicker/persian`
 - `faIRJalali` from `react-day-picker/locale` - use `faIRJalali` from
